test: split entity mapping tests by behavior
