+++ tests/test_entity_mappings_static.py
@@ -1,0 +1,103 @@
+"""Static mapping tests split from test_entity_mappings.py."""
+
+"""Tests for entity_mappings.py helper functions."""
+
+from custom_components.thessla_green_modbus import mappings as em
+from custom_components.thessla_green_modbus.mappings import _infer_icon
+
+# ---------------------------------------------------------------------------
+# _infer_icon (lines 191-201)
+# ---------------------------------------------------------------------------
+
+
+def test_infer_icon_temperature_unit():
+    assert _infer_icon("x", "°C") == "mdi:thermometer"
+
+
+def test_infer_icon_temperature_in_name():
+    assert _infer_icon("room_temperature", None) == "mdi:thermometer"
+
+
+def test_infer_icon_flow_unit_m3h():
+    assert _infer_icon("x", "m³/h") == "mdi:fan"
+
+
+def test_infer_icon_flow_unit_m3h_ascii():
+    assert _infer_icon("x", "m3/h") == "mdi:fan"
+
+
+def test_infer_icon_flow_in_name():
+    assert _infer_icon("supply_flow_rate", None) == "mdi:fan"
+
+
+def test_infer_icon_fan_in_name():
+    assert _infer_icon("fan_speed", None) == "mdi:fan"
+
+
+def test_infer_icon_percentage_unit():
+    assert _infer_icon("x", "%") == "mdi:percent-outline"
+
+
+def test_infer_icon_percentage_in_name():
+    assert _infer_icon("speed_percentage", None) == "mdi:percent-outline"
+
+
+def test_infer_icon_time_unit_min():
+    assert _infer_icon("x", "min") == "mdi:timer"
+
+
+def test_infer_icon_time_unit_s():
+    assert _infer_icon("x", "s") == "mdi:timer"
+
+
+def test_infer_icon_time_unit_h():
+    assert _infer_icon("x", "h") == "mdi:timer"
+
+
+def test_infer_icon_time_in_name():
+    assert _infer_icon("boost_time", None) == "mdi:timer"
+
+
+def test_infer_icon_voltage():
+    assert _infer_icon("x", "V") == "mdi:sine-wave"
+
+
+def test_infer_icon_fallback():
+    assert _infer_icon("unknown_register", None) == "mdi:numeric"
+
+
+def test_infer_icon_fallback_unknown_unit():
+    assert _infer_icon("x", "ppm") == "mdi:numeric"
+
+
+# ---------------------------------------------------------------------------
+# _resolve_entity_id (lines 154, 179-184)
+# ---------------------------------------------------------------------------
+
+
+def test_legacy_entity_id_mapper_removed() -> None:
+    """Legacy entity-id mapper has been removed from public mapping API."""
+    assert not hasattr(em, "map_legacy_entity_id")
+
+
+# ---------------------------------------------------------------------------
+# Phase 6 — _parse_states pure-function tests
+# ---------------------------------------------------------------------------
+
+
+def test_parse_states_valid_entries():
+    """Valid 'value - label' parts are added to state map (line 243)."""
+    from custom_components.thessla_green_modbus.mappings import _parse_states
+
+    states = _parse_states("0 - off; 1 - on")
+    assert states == {"off": 0, "on": 1}
+
+
+def test_parse_states_skips_empty_parts():
+    """Empty parts from consecutive semicolons are skipped (line 237)."""
+    from custom_components.thessla_green_modbus.mappings import _parse_states
+
+    states = _parse_states("0 - off;;1 - on")
+    assert states == {"off": 0, "on": 1}
+
+
